deleted deprecated utils funcs
